Refectored Drawing acivity in three diffrent classes in order to increase modularity and added offline mode
